# perf(encoding): speed up EncodeVarint with io.ByteWriter+hand rolled varintEncode